// react-module/src/openapi/models/machine-calendar-task-calendar-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface MachineCalendarTaskCalendarDto
 */
export interface MachineCalendarTaskCalendarDto {
    /**
     * 
     * @type {number}
     * @memberof MachineCalendarTaskCalendarDto
     */
    id?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskCalendarDto
     */
    startDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskCalendarDto
     */
    startDatePlanned?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskCalendarDto
     */
    endDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskCalendarDto
     */
    endDatePlanned?: any;
    /**
     * 
     * @type {Array&lt;AtomicOperationProductionOrderLocalErpOrderDto&gt;}
     * @memberof MachineCalendarTaskCalendarDto
     */
    operationList?: any;
    /**
     * 
     * @type {OperationTypeBasicDto}
     * @memberof MachineCalendarTaskCalendarDto
     */
    operationType?: any;
    /**
     * 
     * @type {MachineDto}
     * @memberof MachineCalendarTaskCalendarDto
     */
    machine?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskCalendarDto
     */
    planningStatus?: MachineCalendarTaskCalendarDtoPlanningStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskCalendarDto
     */
    productionStatus?: MachineCalendarTaskCalendarDtoProductionStatusEnum;
    /**
     * 
     * @type {ForSavingEstimatedTime}
     * @memberof MachineCalendarTaskCalendarDto
     */
    estimatedTime?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskCalendarDto
     */
    info?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskCalendarDtoPlanningStatusEnum {
    INIT = 'INIT',
    PLANNED = 'PLANNED',
    CONFIRMED = 'CONFIRMED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED'
}
/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskCalendarDtoProductionStatusEnum {
    INIT = 'INIT',
    INPROGRESS = 'IN_PROGRESS',
    POSTPONED = 'POSTPONED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED',
    FAULT = 'FAULT'
}